Moved some scripts to proper folders
