+++ Bank/BackEnd/Databases/CreateTables.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS "Teller" (
 	"ID"	INTEGER NOT NULL UNIQUE,
 	"Name"	TEXT NOT NULL,
 	"LastName"	TEXT NOT NULL,
-	"User"	TEXT NOT NULL UNIQUE,
+	"Username"	TEXT NOT NULL UNIQUE,
 	"Password"	TEXT NOT NULL,
 	PRIMARY KEY("ID" AUTOINCREMENT)
 );
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS "University" (
 	"ID"	INTEGER NOT NULL UNIQUE,
 	"Name"	TEXT NOT NULL UNIQUE,
 	"DepositAmount"	NUMERIC NOT NULL,
-	"User"	TEXT NOT NULL UNIQUE,
+	"Username"	TEXT NOT NULL UNIQUE,
 	"Password"	TEXT NOT NULL,
 	"Email"	TEXT NOT NULL UNIQUE,
 	"Phone"	TEXT,
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS "User" (
 	"ID"	INTEGER NOT NULL UNIQUE,
 	"Name"	TEXT NOT NULL,
 	"LastName"	TEXT NOT NULL,
-	"User"	TEXT NOT NULL UNIQUE,
+	"Username"	TEXT NOT NULL UNIQUE,
 	"Password"	TEXT NOT NULL,
 	"Email"	TEXT NOT NULL,
 	"Phone"	TEXT,
