--- Bank/BackEnd/endpoints/database.js
@@ -1,14 +1,0 @@
-const sqlite3 = require("sqlite3").verbose();
-
-const DB_PATH = "./Bank.db";
-
-const db = new sqlite3.Database(DB_PATH, (err) => {
-    if (err) {
-        console.error(err.message);
-        throw err;
-    }
-    else
-        console.log("Connected to the SQLite database.");
-});
-
-module.exports = db;
