--- Bank/BackEnd/endpoints/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "estudent_enrollement",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.18.1",
-    "md5": "^2.3.0",
-    "nodemon": "^2.0.19",
-    "sqlite3": "^5.0.8"
-  }
-}
