+++ Bank/BackEnd/package.json
@@ -6,9 +6,10 @@
   "scripts": {
     "start": "node server.js",
     "start-server": "node server.js",
+    "dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "juanpavon",
+  "author": "bank",
   "license": "ISC",
   "dependencies": {
     "express": "^4.18.1",
